% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestE.r
\name{best.E.f}
\alias{best.E.f}
\title{Find the best E variable and model to describe another biological variable}
\usage{
best.E.f(x.matrix, y)
}
\arguments{
\item{x.matrix}{a matrix of explanatory (E) variables}

\item{y}{the response variable}
}
\description{
Fits multiple scam models of response~s(explanatory), choose the best scam for each explanatory variable and then
         produces a list of scam objects as well as a vector of deviance explained by each variable. One would likely want an E
         variables that explains more of the variance but you also need to consider the amount of data and the shape of the model
         fitted.

         Presently, the function will fail on the inability to fit a scam model. It needs to be error proofed using tryCatch.
}
\keyword{AIC,}
\keyword{best}
\keyword{constraint,}
\keyword{deviance}
\keyword{explained}
\keyword{fit,}
\keyword{model,}
\keyword{shape}
