% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestE.r
\name{best.model.f}
\alias{best.model.f}
\title{Find the best shape constrained GAM for y~x}
\usage{
best.model.f(x, y, plot = F)
}
\arguments{
\item{x}{the explanatory variable}

\item{y}{the response variable}

\item{plot}{if T then a plot of the best model (lowest AIC) fit is plotted with pertinent statistics in plot titles}
}
\description{
Uses the package scam to fit shape constrained gam models of y~ s(x). Gaussian link function. All the scam shape
             are fitted and the best one is selected. If the smooth term is not significant you may not want to use the model.
             It depends on your hypothesis and how you are using the model and residuals.
}
\examples{
      best.model.f(USArrests$UrbanPop, USArrests$Rape, plot=T)
}
\keyword{AIC}
\keyword{best}
\keyword{constraint,}
\keyword{fit,}
\keyword{model,}
\keyword{shape}
